refactor: extract collection item component
