+++ src/app.tsx
@@ -1,12 +1,11 @@
 import classNames from 'classnames/bind';
-import type { Collection, CollectionItem, EnchantedItem } from '../types';
+import type { Collection } from '../types';
 import styles from './app.module.scss';
+import { CollectionItem } from './components/collection-item/collection-item';
 import { FilterInput } from './components/filter-input/filter-input';
 import { Navigation } from './components/navigation/navigation';
 import { Notification } from './components/notification/notification';
-import { itemsMap } from './constants/items-map';
 import { searchParamsConfig } from './constants/search-params-config';
-import imagesJSON from './data/images.json';
 import { useAppSearchParams } from './hooks/use-app-search-params';
 import { useCollections } from './hooks/use-collections';
 
@@ -57,43 +56,8 @@ function CollectionItemsUi({ collection }: { collection: Collection }) {
   return (
     <div className={cx('collection-items-wrapper')}>
       {collection.items.map((item, index) => (
-        <CollectionItemUi key={`${collection.name}/${index}`} item={item} />
+        <CollectionItem key={`${collection.name}/${index}`} item={item} />
       ))}
-    </div>
-  );
-}
-
-function getTitle({ name, enchant, count }: { name: string; enchant?: number; count?: number }): string {
-  let result = name;
-
-  if (enchant) {
-    result = `+${enchant} ${result}`;
-  }
-
-  return `${result} - ${count ?? 1}шт.`;
-}
-
-function CollectionItemUi({ item }: { item: CollectionItem }) {
-  const singleItem: EnchantedItem = Array.isArray(item) ? item[0] : item;
-  const base64Image: string = imagesJSON[singleItem.id.toString() as keyof typeof imagesJSON];
-
-  const title = getTitle({
-    name: itemsMap.get(singleItem.id)?.name ?? '',
-    enchant: singleItem.enchant,
-    count: singleItem.count,
-  });
-
-  return (
-    <div className={cx('collection-item')} title={title}>
-      {singleItem.enchant && <div className={cx('collection-item-enchant')}>{singleItem.enchant}</div>}
-      {singleItem.count && singleItem.count > 1 && (
-        <div className={cx('collection-item-count')}>{singleItem.count}</div>
-      )}
-      <img
-        className={cx('collection-item-image')}
-        src={`data:image/gif;base64,${base64Image}`}
-        alt={String(singleItem.id)}
-      />
     </div>
   );
 }
