+++ src/components/collection-item/helpers/get-title.ts
@@ -1,0 +1,9 @@
+export function getTitle({ name, enchant, count }: { name: string; enchant?: number; count?: number }): string {
+  let result = name;
+
+  if (enchant) {
+    result = `+${enchant} ${result}`;
+  }
+
+  return `${result} - ${count ?? 1}шт.`;
+}
